refactor: fix installer to not run on mac

+++ uninstall.sh
@@ -1,7 +1,11 @@
 #!/usr/bin/env bash
 
-# Uninstall php using apt
-sudo apt remove php8.0 php8.0-curl -y
+SYSTEM=$3
 
-# Uninstall composer
-sudo rm /usr/local/bin/composer
+if [ "$SYSTEM" != "mac" ]; then
+    # Uninstall php using apt
+    sudo apt remove php8.0 php8.0-curl -y
+
+    # Uninstall composer
+    sudo rm /usr/local/bin/composer
+fi
